feat: configure storybook preview for chakra ui
