+++ packages/cli/src/commands/setup/chakra-ui/tasks/configure-storybook.js
@@ -1,0 +1,39 @@
+import fs from 'fs'
+import path from 'path'
+
+import { getPaths } from '../../../../lib'
+
+/**
+ * @param options {{ force?: boolean }}
+ * @returns {"todo" | "done"}
+ */
+export function checkStorybookStatus({ force }) {
+  const { storybookPreviewConfig } = getPaths().web
+
+  if (fs.existsSync(storybookPreviewConfig)) {
+    if (force) {
+      fs.unlinkSync(storybookPreviewConfig)
+    } else {
+      console.log(
+        `Storybook preview config already exists at ${storybookPreviewConfig}\nUse --force to override existing config.`
+      )
+      return 'done'
+    }
+  }
+
+  return 'todo'
+}
+
+/**
+ * Configure Storybook for Chakra UI by creating a custom preview config
+ */
+export function configureStorybook() {
+  const { storybookPreviewConfig } = getPaths().web
+
+  const storybookPreview = fs.readFileSync(
+    path.join(__dirname, '..', 'templates', 'storybook.preview.js.template'),
+    'utf-8'
+  )
+
+  fs.writeFileSync(storybookPreviewConfig, storybookPreview)
+}
